requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
python-dotenv==1.0.1
